Move internal/engine/errors to pkg/engine/errors to reduce coupling
