+++ internal/engine/actions/remediate/noop/noop.go
@@ -12,7 +12,7 @@ import (
 
 	"google.golang.org/protobuf/reflect/protoreflect"
 
-	enginerr "github.com/mindersec/minder/internal/engine/errors"
+	enginerr "github.com/mindersec/minder/pkg/engine/errors"
 	"github.com/mindersec/minder/internal/engine/interfaces"
 	"github.com/mindersec/minder/pkg/profiles/models"
 )
