+++ internal/engine/eval/jq/jq.go
@@ -13,7 +13,7 @@ import (
 
 	"google.golang.org/protobuf/reflect/protoreflect"
 
-	evalerrors "github.com/mindersec/minder/internal/engine/errors"
+	evalerrors "github.com/mindersec/minder/pkg/engine/errors"
 	"github.com/mindersec/minder/internal/engine/eval/templates"
 	"github.com/mindersec/minder/internal/util"
 	pb "github.com/mindersec/minder/pkg/api/protobuf/go/minder/v1"
