+++ internal/engine/ingester/builtin/builtin_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"google.golang.org/protobuf/reflect/protoreflect"
 
-	evalerrors "github.com/mindersec/minder/internal/engine/errors"
+	evalerrors "github.com/mindersec/minder/pkg/engine/errors"
 	pb "github.com/mindersec/minder/pkg/api/protobuf/go/minder/v1"
 )
