+++ internal/logger/telemetry_store.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/mindersec/minder/internal/engine/actions/alert"
 	"github.com/mindersec/minder/internal/engine/actions/remediate"
-	"github.com/mindersec/minder/internal/engine/errors"
+	"github.com/mindersec/minder/pkg/engine/errors"
 	"github.com/mindersec/minder/internal/engine/interfaces"
 )
