+++ pkg/engine/errors/errors_test.go
@@ -1,3 +1,4 @@
+
 // SPDX-FileCopyrightText: Copyright 2024 The Minder Authors
 // SPDX-License-Identifier: Apache-2.0
